Polish how to play dialog, add sample board animations
Also made the dialog component a HTML `<dialog>` element so that close events can easily be added. In the case of the how to play dialog, the close event will clear all animation intervals. It's a non-modal dialog because the notification elements still need to appear on top of the dialogs if they appear at the same time, and I couldn't find a way to make that work in the top layer.